+++ database/migrations/2018_12_27_120006_create_ratings_table.php
@@ -14,9 +14,9 @@ class CreateRatingsTable extends Migration
     public function up()
     {
         Schema::create('ratings', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('photo_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('photo_id')->unsigned();
             $table->tinyInteger('rating_value')->unsigned();
             $table->timestamps();
