+++ database/migrations/2019_01_11_153241_create_subscribe_table.php
@@ -13,24 +13,24 @@ class CreateSubscribeTable extends Migration
      */
     public function up()
     {
-Schema::table('users', function ($table) {
-    $table->string('stripe_id')->nullable()->collation('utf8mb4_bin');
-    $table->string('card_brand')->nullable();
-    $table->string('card_last_four', 4)->nullable();
-    $table->timestamp('trial_ends_at')->nullable();
-});
+        Schema::table('users', function ($table) {
+            $table->string('stripe_id')->nullable()->collation('utf8mb4_bin');
+            $table->string('card_brand')->nullable();
+            $table->string('card_last_four', 4)->nullable();
+            $table->timestamp('trial_ends_at')->nullable();
+        });
 
-Schema::create('subscriptions', function ($table) {
-    $table->increments('id');
-    $table->unsignedInteger('user_id');
-    $table->string('name');
-    $table->string('stripe_id')->collation('utf8mb4_bin');
-    $table->string('stripe_plan');
-    $table->integer('quantity');
-    $table->timestamp('trial_ends_at')->nullable();
-    $table->timestamp('ends_at')->nullable();
-    $table->timestamps();
-});
+        Schema::create('subscriptions', function ($table) {
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id');
+            $table->string('name');
+            $table->string('stripe_id')->collation('utf8mb4_bin');
+            $table->string('stripe_plan');
+            $table->integer('quantity');
+            $table->timestamp('trial_ends_at')->nullable();
+            $table->timestamp('ends_at')->nullable();
+            $table->timestamps();
+        });
     }
 
     /**
